[feat] (주수아) 카드 추천 결제 1차 완성

+++ back/moapay/src/main/java/com/moa/moapay/domain/card/model/BenefitInfo.java
@@ -8,8 +8,8 @@ import lombok.*;
 @AllArgsConstructor
 @NoArgsConstructor
 public class BenefitInfo {
-    private long totalDiscount;
-    private long totalPoint;
-    private long totalCashback;
-    private long totalBenefit;
+    private double totalDiscount;
+    private double totalPoint;
+    private double totalCashback;
+    private double totalBenefit;
 }
